app-ui/api/requirements.txt
Flask==2.2.5
azure-identity==1.13.0
langchain==0.0.187
openai==1.6.1
tiktoken==0.4.0
azure-search-documents==11.4.0b6
azure.storage.blob==12.14.1
pypdf==3.9.0
azure-ai-formrecognizer==3.2.1
pymongo==3.12.0
azure-mgmt-cognitiveservices==13.5.0
chardet==5.2.0
azure-ai-textanalytics==5.3.0
msrestazure